requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.7
django-redis==6.0.0
djangorestframework==3.16.1
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
